Revert "ReVerify the amount"
This reverts commit 2370219d6419130bc967e005a461622f44d0d11b.

+++ HRS.API/Contracts/DTOs/RentalOrder/RentalOrderRequestDto.cs
@@ -89,7 +89,7 @@ public class AssignStripeSessionRequest
 public class ApprovePaymentRequest
 {
     public string SessionId { get; set; } = string.Empty;
-    public long? Amount { get; set; }
+    public int? Amount { get; set; }
 }
 
 public class UpdateItemRequestDto
